+++ src/elements/index.js
@@ -16,3 +16,4 @@ export { default as Pagination } from './Pagination';
 export { default as Label } from './Label';
 export { default as Loader } from './Loader';
 export { default as Dropdown } from './Dropdown';
+export { default as PillLabel } from './PillLabel';